requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jedi==0.19.2
packaging==24.2
parso==0.8.4
pudb==2024.1.3
Pygments==2.19.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_ex
